feat: updated integration test cases
